+++ src/gc_monitor/__init__.py
@@ -36,7 +36,7 @@ Example usage:
     monitor.stop_monitoring()
 """
 
-__version__ = "0.1.0"
+__version__ = "0.3.3"
 __author__ = "Akshat Kotpalliwar alias IntegerAlex"
 __license__ = "LGPL-2.1-only"